LINUX-7026 - Module errata support for OL8
Update server errata calculations to handle Oracle module errata.

Spacewalk cannot parse or generate modularity metadata. This means
that it treats all packages in a channel with modules and streams as
if they were in a traditional channel. The result of this is that
Spacewalk erroneously reports that errata from other module:stream, or
even the plain channel, are applicable to packages from another
module:stream.

All packages released by Oracle have an errata associated with
them. Module:stream errata have the module:stream name in the errata
synopsis. We use this to allow us to determine the currently installed
streams and restrict errata recommendations for module:stream to only
the installed module:stream.

A critical caveat is that this only works for errata that have the
module:stream in the synopsis. We cannot guarantee this is the case
for other vendors, and therefore can only state that this should work
for Oracle errata.

Another caveat is that this only works for packages in a module:stream
that uses packages explicitly marked as a module packages. Some
module:stream exist that use normal packages. This enhancement cannot
accommodate that scenario. There will be edge cases where a package
installed from the plain channel or a module:stream that uses normal
packages will report as vulnerable to errata from a newer
module:stream. There is insufficient data to avoid this scenario.

Jira: LINUX-7026

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>

Revert "1581744 - disable errata updates when modules are present"

This reverts commit a821379f5f881cbe8b044b3037f34add8f89a251.

Orabug: 31665403
Jira: LINUX-7026

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>
Reviewed-by: Darren Kenny <darren.kenny@oracle.com>