Fix remote command execution erroneously reporting failure.
Successful remote command execution was being reported by SW as a
failure because cx_Oracle was incorrectly handling the CLOB objects
used to store the result in the DB.

Python was reporting "'list' object has no attribute
'write'". Separating declaration and assignment avoids missing
__item__ errors by enforcing correct dereferencing.

This error was introduced when the Instant Client version got
bumped. It caused cx_Oracle to return CLOBS as lists of LOB objects
rather than a direct pointer to the CLOB.

Adding an index into the nested array fixes the issue.

Orabug: 31589572

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: lukas.lipinsky@oracle.com