LINUX-7708 - Copy modules.yaml metadata file when cloning channels.
When cloning a channel the reference to the modules.yaml file was just
a pointer to the original channel's file. When the original file was
updated with new modules it would cause Spacewalk to report modules
and packages for the clone channel that aren't present in the clone
channel. Attempting to install the new modules and packages from the
clone channel would fail.

Cloning a channel with modules will now copy the modules.yaml file,
rather than pointing at the original. This means the clone channel and
modules list will never go out of sync.

This change is implemented in ChannelManager and exposed to the Web UI
controller layer and via the XMLRPC channels.software API.

Jira: LINUX-7708

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>

Make reposync set rhn/modules group writable

Copying modules.yaml files when cloning a channel requires that
rhn/modules is group-writable. Reposync is responsible for creating
this path, so have it set the permissions too.

Jira: LINUX-7708

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>

Make modules metadata folder group-writable on spacewalk-backend upgrade

Copying the modules.yaml file when cloning channels requires that
rhn/modules path is group-writable. Have rpm determine the path from
rhn.conf and set the permissions.

Jira: LINUX-7708

Signed-off-by: Laurence Rochfort <laurence.rochfort@oracle.com>
Reviewed-by: Alex Burmashev <alexander.burmashev@oracle.com>